+++ tools/cmake/samples/cflags/add_compile_definitions.cmake
@@ -1,0 +1,1 @@
+add_compile_definitions(DEF_1=1)