+++ tools/cmake/samples/cflags/add_definitions.cmake
@@ -1,0 +1,1 @@
+add_definitions(-DDEF_2=2)